+++ client/src/components/common/markdown/index.tsx
@@ -1,15 +1,13 @@
 import React from "react";
 import ReactMarkdown from "react-markdown";
-import { Code } from "./renderers";
+import {Code} from "./renderers";
 
 interface IProps {
     value?: string
 }
 
-const Markdown: React.FC<IProps> = ({ value }) => {
-    return <ReactMarkdown source={value} renderers={{
-        code: Code
-    }} />;
-};
+const Markdown: React.FC<IProps> = ({value}) => <ReactMarkdown source={value} renderers={{
+  code: Code,
+}} />;
 
-export default Markdown
+export default Markdown;
