+++ client/src/components/common/markdown/renderers/code.tsx
@@ -1,15 +1,22 @@
 import React from "react";
+import Youtube from "../extensions/Youtube";
 
 interface IProps {
     language: string | undefined,
-    value: string | undefined
+    value: string
 }
 
-const Code: React.FC<IProps> = ({ language, value }) => (
+const Code: React.FC<IProps> = ({language, value}) => {
+  if (language === "youtube") {
+    return (<Youtube code={value} />);
+  }
+
+  return (
     <pre>
-        <code>{value}</code>
+      <code>{value}</code>
     </pre>
-);
+  );
+};
 
 
 export default Code;
