+++ client/src/components/common/markdown/renderers/index.ts
@@ -1,4 +1,4 @@
-import Code from "./code"
+import Code from "./Code";
 
 export {
     Code
