+++ client/src/components/editor/Preview.tsx
@@ -5,7 +5,7 @@ interface IProps {
   content: string
 }
 
-const Preview: React.FC<IProps> = ({ content }) => (
+const Preview: React.FC<IProps> = ({content}) => (
   <div>
     <Markdown value={content} />
   </div>
